Add TokenAuthorizerMixin
This fixes issue #218

All options are the same as the old authorizer but uses the new syntax.
The Mixin will import DataAdapterMixin and session for you.

I was unsure on tests so would appreciate anyone pushing a commit or two
to help with that.

Added a blurb to the README but it could use some help from someone more
articulate then I.