Fix Return For Token Authenticator
I pinpointed an issue I was experiencing with the Token Authenticator
and it _always_ returning `undefined`. This seemed to be something to do
with unwrapping Promise values. The `makeRequest` method was
unnecessarily wrapping `fetch`. In removing this wrapping and leaning
more on built in Promise semantics the problem went away.